+++ app/web.py
@@ -8,6 +8,7 @@ from app import app, Invitations, Settings, VERSION, get_locale
 from app.plex import *
 from flask_babel import _
 import threading
+import requests
 
 
 @app.route('/')
@@ -59,9 +60,32 @@ def connect():
     elif Settings.get(key="server_type").value == "jellyfin":
         return render_template("signup-jellyfin.html", code=code)
 
+def ombi_RunUserImporter(name):
+    if not Settings.get_or_none(Settings.key == "overseerr_url"):
+        return
+    if not Settings.get_or_none(Settings.key == "ombi_api_key"):
+        return
+
+    overseerr_url = Settings.get_or_none(Settings.key == "overseerr_url").value
+    ombi_api_key = Settings.get_or_none(Settings.key == "ombi_api_key").value
+    headers = {
+        "ApiKey": ombi_api_key,
+    }
+    response = requests.post(
+        f"{overseerr_url}/api/v1/Job/{name}UserImporter/", headers=headers)
+    logging.info(f"POST {overseerr_url}/api/v1/Job/{name}UserImporter/ - {str(response.status_code)}")
+
+    return response
+
+def ombi_RunAllUserImporters():
+    #ombi_RunUserImporter('plex')
+    #ombi_RunUserImporter('emby')
+    return ombi_RunUserImporter('jellyfin')
 
 @app.route('/setup', methods=["GET"])
 def setup():
+    ombi_RunAllUserImporters()
+
     resp = make_response(render_template(
         "wizard.html", server_type=Settings.get(Settings.key == "server_type").value))
     resp.set_cookie('current', "0")
